spenso::iterators

Trait AbstractFiber

pub trait AbstractFiber<Out: AbstractFiberIndex>: Index<usize, Output = Out> {
    type Repr: RepName;

    // Required methods
    fn strides(&self) -> Vec<usize>;
    fn shape(&self) -> Vec<Dimension>;
    fn reps(&self) -> Vec<Representation<Self::Repr>>;
    fn order(&self) -> usize;
    fn single(&self) -> Option<usize>;
    fn bitvec(&self) -> SubSet<SlotIndex>;
}

Abstract trait defining fiber properties

This trait defines the properties of a fiber, which is a slice through a tensor along specific dimensions.

Required Associated Types

type Repr: RepName

Representation type for the fiber

Required Methods

fn strides(&self) -> Vec<usize>

Returns the strides of the fiber

fn shape(&self) -> Vec<Dimension>

Returns the shape of the fiber

fn reps(&self) -> Vec<Representation<Self::Repr>>

Returns the representations of the fiber dimensions

fn order(&self) -> usize

Returns the order (number of dimensions) of the fiber

fn single(&self) -> Option<usize>

Returns the single varying dimension if there is one

fn bitvec(&self) -> SubSet<SlotIndex>

Returns a bitvector indicating free (true) and fixed (false) dimensions