spenso::iterators

Trait AbstractFiberIndex

pub trait AbstractFiberIndex {
    // Required method
    fn is_free(&self) -> bool;

    // Provided method
    fn is_fixed(&self) -> bool { ... }
}

Abstract trait for fiber indices

This trait defines methods for determining whether an index is free or fixed, which is fundamental to how fibers are defined and traversed.

Required Methods

fn is_free(&self) -> bool

Returns true if the index is free (varying during iteration)

Provided Methods

fn is_fixed(&self) -> bool

Returns true if the index is fixed (constant during iteration)

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".