spenso::iterators

Trait FiberIteratorItem

pub trait FiberIteratorItem {
    type OtherData;

    // Required methods
    fn flat_idx(&self) -> FlatIndex;
    fn other_data(self) -> Self::OtherData;
}

Trait for items yielded by fiber iterators

Defines the interface for items returned by fiber iterators, including their flat index representation and any additional data they carry.

Required Associated Types

type OtherData

Additional data carried by the iterator item

Required Methods

fn flat_idx(&self) -> FlatIndex

Returns the flat index of the item

fn other_data(self) -> Self::OtherData

Extracts additional data from the item