spenso::iterators

Trait IteratableTensor

pub trait IteratableTensor:
    HasStructure
    + Sized
    + TensorStructure {
    type Data<'a>
       where Self: 'a;

    // Required methods
    fn iter_expanded(
        &self,
    ) -> impl Iterator<Item = (ExpandedIndex, Self::Data<'_>)>;
    fn iter_flat(&self) -> impl Iterator<Item = (FlatIndex, Self::Data<'_>)>;

    // Provided methods
    fn fiber<'a>(&'a self, fiber_data: FiberData<'_>) -> Fiber<'a, Self> { ... }
    fn fiber_mut<'a>(
        &'a mut self,
        fiber_data: FiberData<'_>,
    ) -> FiberMut<'a, Self> { ... }
    fn fiber_class<'a>(
        &'a self,
        fiber_data: FiberData<'_>,
    ) -> FiberClass<'a, Self> { ... }
    fn fiber_class_mut<'a>(
        &'a mut self,
        fiber_data: FiberData<'_>,
    ) -> FiberClassMut<'a, Self> { ... }
}

Trait for tensor types that can be iterated

Defines the interface for tensor types that support various forms of iteration.

Required Associated Types

type Data<'a> where Self: 'a

Type of data yielded by iterators over this tensor

Required Methods

fn iter_expanded(&self) -> impl Iterator<Item = (ExpandedIndex, Self::Data<'_>)>

Returns an iterator over expanded indices and data

fn iter_flat(&self) -> impl Iterator<Item = (FlatIndex, Self::Data<'_>)>

Returns an iterator over flat indices and data

Provided Methods

fn fiber<'a>(&'a self, fiber_data: FiberData<'_>) -> Fiber<'a, Self>

Creates a fiber from the tensor

Arguments

fiber_data - Data specifying which dimensions are fixed/free

fn fiber_mut<'a>(&'a mut self, fiber_data: FiberData<'_>) -> FiberMut<'a, Self>

Creates a mutable fiber from the tensor

Arguments

fiber_data - Data specifying which dimensions are fixed/free

fn fiber_class<'a>(&'a self, fiber_data: FiberData<'_>) -> FiberClass<'a, Self>

Creates a fiber class from the tensor

Arguments

fiber_data - Data specifying which dimensions are fixed/free

fn fiber_class_mut<'a>( &'a mut self, fiber_data: FiberData<'_>, ) -> FiberClassMut<'a, Self>

Creates a mutable fiber class from the tensor

Arguments

fiber_data - Data specifying which dimensions are fixed/free

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.