spenso::iterators

Trait IteratesAlongFibers

pub trait IteratesAlongFibers<R: RepName>:
    Iterator
    + ShiftableIterator
    + ResetableIterator {
    // Required methods
    fn new<I, J>(fiber: &I, conj: bool) -> Self
       where I: AbstractFiber<J, Repr = R>,
             J: AbstractFiberIndex,
             Self: Sized;
    fn new_paired_conjugates<I, J>(fiber: &I) -> (Self, Self)
       where I: AbstractFiber<J, Repr = R>,
             J: AbstractFiberIndex,
             Self: Sized;
}

Trait for iterators that traverse along fibers

Defines the core operations for iterators that move along tensor fibers, with support for resetting and creating paired conjugate iterators.

Required Methods

fn new<I, J>(fiber: &I, conj: bool) -> Self
where I: AbstractFiber<J, Repr = R>, J: AbstractFiberIndex, Self: Sized,

Create a new iterator from a fiber

Arguments

fiber - The fiber to iterate along

conj - Whether to use conjugate iteration

fn new_paired_conjugates<I, J>(fiber: &I) -> (Self, Self)
where I: AbstractFiber<J, Repr = R>, J: AbstractFiberIndex, Self: Sized,

Create a pair of conjugate iterators from a fiber

Arguments

fiber - The fiber to iterate along