spenso

Macro chain

macro_rules! chain {
    ($start:expr, $end:expr $(,)?) => { ... };
    ($start:expr, $end:expr; $factors:expr $(,)?) => { ... };
    ($start:expr, $end:expr $(, $factor:expr)+; $factors:expr $(,)?) => { ... };
    ($start:expr, $end:expr $(, $factor:expr)* $(,)?) => { ... };
}

Builds a symbolic open chain with a start slot, an end slot, and any number of factors.

Arguments are converted through IntoAtom, so callers can pass Spenso slots, atoms, atom views, or symbols. This macro is only the variadic surface for SPENSO_TAG.chain; it does not simplify or normalize the expression.

Examples

use spenso::{chain, slot};

let expr = chain!(
    slot!(bis4, a),
    slot!(bis4, b),
    gamma_mu,
    gamma_nu,
);

let factors = vec![gamma_mu, gamma_nu];
let expr = chain!(slot!(bis4, a), slot!(bis4, b); factors);