spenso

Macro slot

macro_rules! slot {
    ($rep:expr, $index:ident) => { ... };
    ($rep:expr, $index:expr) => { ... };
}

Creates an abstract-index slot from a representation and an index.

The slot index is fixed to Spenso’s symbolic AbstractIndex, which avoids type-inference ambiguity when passing Symbols. The second argument can be an identifier, expanded through s!, an integer index, or an arbitrary expression convertible into AbstractIndex.

Examples

use spenso::{s, slot};
use spenso::structure::representation::{Minkowski, RepName};

let mink4 = Minkowski {}.new_rep(4);
let mu = slot!(mink4, mu);
let nu = slot!(mink4, s!(nu));
let one = slot!(mink4, 1);