spenso

Macro trace

macro_rules! trace {
    ($rep:expr $(,)?) => { ... };
    ($rep:expr; $factors:expr $(,)?) => { ... };
    ($rep:expr $(, $factor:expr)+; $factors:expr $(,)?) => { ... };
    ($rep:expr $(, $factor:expr)* $(,)?) => { ... };
}

Builds a symbolic trace with a representation and any number of factors.

The representation and factors are converted through IntoAtom. Empty traces are emitted as trace(rep). Non-empty traces are emitted in the canonical cyclic form trace(rep, cyclic(factors...)).

Examples

use spenso::{cyclic, trace};

let expr = trace!(
    &cof_nc,
    color_t_a,
    color_t_b,
);
assert_eq!(expr, trace!(&cof_nc, cyclic!(color_t_a, color_t_b)));

let factors = vec![color_t_a, color_t_b];
let expr = trace!(&cof_nc; factors);