spenso::network::graph

Type Alias NetworkHedgeGraph

pub type NetworkHedgeGraph<K, FK, Aind> = HedgeGraph<NetworkEdge<Aind>, NetworkNode<K, FK, Aind>, NoData, NetworkNodeStore<K, FK, Aind>>;

Aliased Type

pub struct NetworkHedgeGraph<K, FK, Aind> {
    pub node_store: Forest<NetworkNode<K, FK, Aind>, ParentChildStore<()>>,
    /* private fields */
}

Fields
node_store: Forest<NetworkNode<K, FK, Aind>, ParentChildStore<()>>

Storage for all nodes in the graph, including their data (V) and information about the half-edges incident to them. The specific implementation is determined by the S type parameter.