spenso::network

Module parsing

Structs

ParseSettings

ParseState

SchoonschipExpansionMode

Enums

ShorthandParsing

StrictTensorFilter

StructureInferenceMode

Chooses how tensor structure is inferred from symbolic syntax.

Traits

AtomStructureExt

NetworkParse

StructureFromAtom

TensorFromExpression

Builds a tensor leaf from an opaque expression once its structure is known.

TensorLibraryFor

A tensor library whose stored tensor can be indexed into T.

Type Aliases

ParamNet

ShadowedStructure