spenso::network::parsing

Type Alias ParamNet

pub type ParamNet<Aind> = Network<NetworkStore<ParamTensor<ShadowedStructure<Aind>>, Atom>, DummyKey, Symbol, Aind>;

Aliased Type

pub struct ParamNet<Aind> {
    pub graph: NetworkGraph<DummyKey, Symbol, Aind>,
    pub store: NetworkStore<ParamTensor<NamedStructure<Symbol, Vec<Atom>, LibraryRep, Aind>>, Atom>,
    pub state: NetworkState,
}

Fields
graph: NetworkGraph<DummyKey, Symbol, Aind>

Expression and contraction topology whose leaves reference stored values.

store: NetworkStore<ParamTensor<NamedStructure<Symbol, Vec<Atom>, LibraryRep, Aind>>, Atom>

Backing tensor and scalar storage referenced by graph leaves.

state: NetworkState

Classification of the network result as scalar or tensor, including self-duality.

Implementations

impl<Aind: AbsInd + DummyAind + ParseableAind + 'static> ParamNet<Aind>

pub fn simple_execute(&mut self)