spenso::shadowing

Trait Shadowable

pub trait Shadowable: HasStructure<Structure: TensorStructure + HasName<Name: IntoSymbol, Args: IntoArgs> + Clone + Sized> + Sized
where
    <<Self::Structure as TensorStructure>::Slot as IsAbstractSlot>::Aind: ParseableAind,
{
    // Provided methods
    fn expanded_shadow(&self) -> Result<DenseTensor<Atom, Self::Structure>> { ... }
    fn expanded_shadow_perm(
        &self,
        perm: &Permutation,
    ) -> Result<DenseTensor<Atom, Self::Structure>> { ... }
    fn flat_shadow(&self) -> Result<DenseTensor<Atom, Self::Structure>> { ... }
    fn shadow<T>(
        &self,
        index_to_atom: impl Fn(&Self::Structure, FlatIndex) -> T,
    ) -> Result<DenseTensor<Atom, Self::Structure>>
       where T: TensorCoefficient { ... }
}

Trait that enables shadowing of a tensor

This creates a dense tensor of atoms, where the atoms are the expanded indices of the tensor, with the global name as the name of the labels.

Provided Methods

fn expanded_shadow(&self) -> Result<DenseTensor<Atom, Self::Structure>>

fn expanded_shadow_perm( &self, perm: &Permutation, ) -> Result<DenseTensor<Atom, Self::Structure>>

fn flat_shadow(&self) -> Result<DenseTensor<Atom, Self::Structure>>

fn shadow<T>( &self, index_to_atom: impl Fn(&Self::Structure, FlatIndex) -> T, ) -> Result<DenseTensor<Atom, Self::Structure>>
where T: TensorCoefficient,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.