spenso::shadowing

Trait TensorCollectExt

pub trait TensorCollectExt {

    // Required methods
    fn collect_tensors(&self) -> Atom;
    fn collect_tagged_tensors(&self) -> Atom;
    fn collect_rep(&self, rep: LibraryRep) -> Atom;
    fn collect_rep_with_map<F: FnMut(AtomView<'_>, &Context, &mut Settable<'_, Atom>)>(
        &self,
        rep: LibraryRep,
        map: F,
    ) -> Atom;
    fn collect_reps<const N: usize>(&self, reps: [LibraryRep; N]) -> Atom;
    fn collect_metrics(&self) -> Atom;
    fn collect_chains_and_traces(&self) -> Atom;
    fn expand_tensors(&self) -> Atom;
    fn expand_tagged_tensors(&self) -> Atom;
    fn expand_rep(&self, rep: LibraryRep) -> Atom;
    fn expand_rep_with_map<F: FnMut(AtomView<'_>, &Context, &mut Settable<'_, Atom>)>(
        &self,
        rep: LibraryRep,
        map: F,
    ) -> Atom;
    fn expand_reps<const N: usize>(&self, reps: [LibraryRep; N]) -> Atom;
    fn expand_metrics(&self) -> Atom;
    fn expand_chains_and_traces(&self) -> Atom;

}

Collect tensor factors without full expression expansion.

Required Methods

fn collect_tensors(&self) -> Atom

Collect common tensor leaves by temporarily wrapping them in spenso::collect(...).

fn collect_tagged_tensors(&self) -> Atom

Collect common tagged tensor leaves by temporarily wrapping them in spenso::collect(...).

fn collect_rep(&self, rep: LibraryRep) -> Atom

Collect common tensor leaves that contain rep as one of their slot representations.

fn collect_rep_with_map<F: FnMut(AtomView<'_>, &Context, &mut Settable<'_, Atom>)>( &self, rep: LibraryRep, map: F, ) -> Atom

Collect common tensor leaves that contain rep as one of their slot representations, using a custom map function.

fn collect_reps<const N: usize>(&self, reps: [LibraryRep; N]) -> Atom

Collect common tensor leaves that contain any of the reps as one of their slot representations.

fn collect_metrics(&self) -> Atom

Collect common metric tensors.

fn collect_chains_and_traces(&self) -> Atom

Collect common chain and trace tensors.

fn expand_tensors(&self) -> Atom

Expand common tensor leaves by temporarily wrapping them in spenso::collect(...).

fn expand_tagged_tensors(&self) -> Atom

Expand common tagged tensor leaves by temporarily wrapping them in spenso::collect(...).

fn expand_rep(&self, rep: LibraryRep) -> Atom

Expand common tensor leaves that contain rep as one of their slot representations.

fn expand_rep_with_map<F: FnMut(AtomView<'_>, &Context, &mut Settable<'_, Atom>)>( &self, rep: LibraryRep, map: F, ) -> Atom

Expand common tensor leaves that contain rep as one of their slot representations, using a custom map function.

fn expand_reps<const N: usize>(&self, reps: [LibraryRep; N]) -> Atom

Expand common tensor leaves that contain any of the reps as one of their slot representations.

fn expand_metrics(&self) -> Atom

Expand common metric tensors.

fn expand_chains_and_traces(&self) -> Atom

Expand common chain and trace tensors.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.