spenso

Module structure

Re-exports

pub use indexless::IndexLess;

pub use indexless::IndexlessNamedStructure;

pub use named::NamedStructure;

pub use permuted::PermutedStructure;

pub use ordered::OrderedStructure;

pub use smart_shadow::SmartShadowStructure;

Modules

abstract_index

concrete_index

dimension

indexless

named

ordered

permuted

representation

slot

smart_shadow

Structs

SlotIndex

TensorShell

Enums

MergeInfo

An enum to represent information about a clean split in merged structures

StructureError

Traits

CastStructure

HasName

HasStructure

ScalarStructure

ScalarTensor

StructureContract

A trait for a structure that can be traced and merged, during a contraction.

TensorStructure

Ordered structural metadata for a tensor’s external coordinates.

ToSymbolic

TracksCount