spenso::structure::representation

Struct Representation

pub struct Representation<T: RepName> {
    pub rep: T,
    pub dim: Dimension,
}

Fields
rep: Tdim: Dimension
Implementations

impl<T: RepName> Representation<T>

pub fn id<'a, It: Into<AtomOrView<'a>>>(&self, a: It, b: It) -> Atom

An atom representing the identity tensor with aind a, and b. a is dualized, b is not.

pub fn inner_product<'a, It: Into<AtomOrView<'a>>>(&self, a: It, b: It) -> Atom

An atom representing the identity tensor with aind a, and b. a is dualized, b is not.

pub fn base(self) -> Representation<T::Base>

pub fn g<'a, It: Into<AtomOrView<'a>>>(&self, a: It, b: It) -> Atom

An atom representing the metric tensor with aind a, and b.

pub fn flat<'a, It: Into<AtomOrView<'a>>>(&self, a: It, b: It) -> Atom

An atom representing the musical isomorphism tensor with aind a, and b.

pub fn to_lib(self) -> Representation<LibraryRep>

pub fn to_dummy(self) -> Representation<Dummy>

pub fn dot(&self) -> String

pub fn pattern<'a, A: Into<AtomOrView<'a>>>(&self, aind: A) -> Atom

pub fn to_pattern_wrapped(&self, aind: Symbol) -> Atom

impl<T: RepName> Representation<T>

pub fn matches(&self, other: &Representation<T::Dual>) -> bool

pub fn match_cmp(&self, other: &Representation<T::Dual>) -> Ordering

pub fn to_symbolic(&self, args: impl IntoIterator<Item = Atom>) -> Atom

yields a function builder for the representation, adding a first variable: the dimension.

pub fn dual(self) -> Representation<T::Dual>

pub fn cast<U: RepName + From<T>>(self) -> Representation<U>

pub fn is_neg(&self, i: usize) -> bool

pub fn slot<Aind, A: Into<Aind>>(&self, aind: A) -> Slot<T, Aind>

pub fn negative(&self) -> Result<Vec<bool>>

gives the vector of booleans, saying which concrete index along a Dimension/Abstract Index should have a minus sign during contraction.