spenso::structure::representation

Trait RepName

pub trait RepName:
    Copy
    + Clone
    + Debug
    + PartialEq
    + Eq
    + Hash
    + Display
    + Ord
    + Into<LibraryRep> {
    type Dual: RepName<Dual = Self, Base = Self::Base>;
    type Base: RepName;

    // Required methods
    fn from_library_rep(rep: LibraryRep) -> Result<Self, RepresentationError>;
    fn orientation(self) -> Orientation;
    fn dual(self) -> Self::Dual;
    fn is_dual(self) -> bool;
    fn base(&self) -> Self::Base;
    fn is_base(&self) -> bool;
    fn matches(&self, other: &Self::Dual) -> bool;

    // Provided methods
    fn is_dummy(self) -> bool { ... }
    fn is_self_dual(&self) -> bool { ... }
    fn match_cmp(&self, _other: &Self::Dual) -> Ordering { ... }
    fn try_from_symbol(
        sym: Symbol,
        aind: Symbol,
    ) -> Result<Self, RepresentationError> { ... }
    fn try_from_symbol_coerced(sym: Symbol) -> Result<Self, RepresentationError> { ... }
    fn is_neg(self, _i: usize) -> bool { ... }
    fn to_symbolic<'a, It: Into<AtomOrView<'a>>>(
        &self,
        args: impl IntoIterator<Item = It>,
    ) -> Atom { ... }
    fn id_atom<'a, It: Into<AtomOrView<'a>>>(
        &self,
        a: impl IntoIterator<Item = It>,
        b: impl IntoIterator<Item = It>,
    ) -> Atom { ... }
    fn metric_atom<'a, It: Into<AtomOrView<'a>>>(
        &self,
        a: impl IntoIterator<Item = It>,
        b: impl IntoIterator<Item = It>,
    ) -> Atom { ... }
    fn new_slot<Aind, D: Into<Dimension>, A: Into<Aind>>(
        self,
        dim: D,
        aind: A,
    ) -> Slot<Self, Aind>
       where Self: Sized { ... }
    fn new_rep<D: Into<Dimension>>(&self, dim: D) -> Representation<Self>
       where Self: Sized { ... }

}

Required Associated Types

type Dual: RepName<Dual = Self, Base = Self::Base>

type Base: RepName

Required Methods

fn from_library_rep(rep: LibraryRep) -> Result<Self, RepresentationError>

fn orientation(self) -> Orientation

fn dual(self) -> Self::Dual

fn is_dual(self) -> bool

fn base(&self) -> Self::Base

fn is_base(&self) -> bool

fn matches(&self, other: &Self::Dual) -> bool

Provided Methods

fn is_dummy(self) -> bool

fn is_self_dual(&self) -> bool

fn match_cmp(&self, _other: &Self::Dual) -> Ordering

fn try_from_symbol( sym: Symbol, aind: Symbol, ) -> Result<Self, RepresentationError>

fn try_from_symbol_coerced(sym: Symbol) -> Result<Self, RepresentationError>

fn is_neg(self, _i: usize) -> bool

for the given concrete index, says whether it should have a minus sign during contraction

For example, see Representation::negative.

fn to_symbolic<'a, It: Into<AtomOrView<'a>>>( &self, args: impl IntoIterator<Item = It>, ) -> Atom

yields a function builder for the representation, adding a first variable: the dimension.

fn id_atom<'a, It: Into<AtomOrView<'a>>>( &self, a: impl IntoIterator<Item = It>, b: impl IntoIterator<Item = It>, ) -> Atom

An atom representing the identity function for that representation. a is dualized b is not

fn metric_atom<'a, It: Into<AtomOrView<'a>>>( &self, a: impl IntoIterator<Item = It>, b: impl IntoIterator<Item = It>, ) -> Atom

yields a function builder for the representation, adding a first variable: the dimension.

fn new_slot<Aind, D: Into<Dimension>, A: Into<Aind>>( self, dim: D, aind: A, ) -> Slot<Self, Aind>
where Self: Sized,

fn new_rep<D: Into<Dimension>>(&self, dim: D) -> Representation<Self>
where Self: Sized,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".