spenso::structure::slot

Trait IsAbstractSlot

pub trait IsAbstractSlot:
    Copy
    + PartialEq
    + Eq
    + Debug
    + Clone
    + Hash
    + Ord
    + Display {
    type Aind: AbsInd;
    type R: RepName;

    // Required methods
    fn reindex(self, id: Self::Aind) -> Self;
    fn dim(&self) -> Dimension;
    fn aind(&self) -> Self::Aind;
    fn set_aind(&mut self, aind: Self::Aind);
    fn rep_name(&self) -> Self::R;
    fn to_atom(&self) -> Atom
       where Self::Aind: ParseableAind;
    fn to_symbolic_wrapped(&self) -> Atom
       where Self::Aind: ParseableAind;

    // Provided methods
    fn to_dummy_rep(&self) -> LibrarySlot<Self::Aind> { ... }
    fn to_dummy_ind(&self) -> LibrarySlot<Self::Aind>
       where Self::Aind: DummyAind { ... }
    fn to_lib(&self) -> LibrarySlot<Self::Aind> { ... }
    fn rep(&self) -> Representation<Self::R> { ... }
}

Required Associated Types

type Aind: AbsInd

type R: RepName

Required Methods

fn reindex(self, id: Self::Aind) -> Self

fn dim(&self) -> Dimension

fn aind(&self) -> Self::Aind

fn set_aind(&mut self, aind: Self::Aind)

fn rep_name(&self) -> Self::R

fn to_atom(&self) -> Atom
where Self::Aind: ParseableAind,

using the function builder of the representation add the abstract index as an argument, and finish it to an Atom.

fn to_symbolic_wrapped(&self) -> Atom
where Self::Aind: ParseableAind,

Provided Methods

fn to_dummy_rep(&self) -> LibrarySlot<Self::Aind>

fn to_dummy_ind(&self) -> LibrarySlot<Self::Aind>
where Self::Aind: DummyAind,

fn to_lib(&self) -> LibrarySlot<Self::Aind>

fn rep(&self) -> Representation<Self::R>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.