spenso::structure

Trait StructureContract

pub trait StructureContract: Sized {
    // Required methods
    fn trace(&mut self, i: usize, j: usize);
    fn trace_out(&mut self);
    fn merge(
        &self,
        other: &Self,
    ) -> Result<(Self, SubSet<SlotIndex>, SubSet<SlotIndex>, MergeInfo), StructureError>;
}

A trait for a structure that can be traced and merged, during a contraction.

Required Methods

fn trace(&mut self, i: usize, j: usize)

fn trace_out(&mut self)

fn merge( &self, other: &Self, ) -> Result<(Self, SubSet<SlotIndex>, SubSet<SlotIndex>, MergeInfo), StructureError>

Merges the two structures, yielding a new ordered structure.

Assumes both are ordered structures

Errors if self or other contained non traced out indices

Otherwise returns a new structure without any common indices, the positions of the common indices self, and the positions in other, along with a MergeInfo specifying whether the merger is clean or not.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.