spenso::structure

Trait TensorStructure

pub trait TensorStructure {
    type Slot: IsAbstractSlot + DualSlotTo<Dual = Self::Slot>;
    type Indexed: TensorStructure<Indexed = Self::Indexed, Slot = Self::Slot>;

    // Required methods
    fn reindex(
        self,
        indices: &[<Self::Slot as IsAbstractSlot>::Aind],
    ) -> Result<PermutedStructure<Self::Indexed>, StructureError>;
    fn dual(self) -> Self;
    fn external_structure_iter(&self) -> impl Iterator<Item = Self::Slot>;
    fn external_dims_iter(&self) -> impl Iterator<Item = Dimension>;
    fn external_reps_iter(
        &self,
    ) -> impl Iterator<Item = Representation<<Self::Slot as IsAbstractSlot>::R>>;
    fn external_indices_iter(
        &self,
    ) -> impl Iterator<Item = <Self::Slot as IsAbstractSlot>::Aind>;
    fn get_aind(
        &self,
        i: impl Into<SlotIndex>,
    ) -> Option<<Self::Slot as IsAbstractSlot>::Aind>;
    fn get_rep(
        &self,
        i: impl Into<SlotIndex>,
    ) -> Option<Representation<<Self::Slot as IsAbstractSlot>::R>>;
    fn get_dim(&self, i: impl Into<SlotIndex>) -> Option<Dimension>;
    fn get_slot(&self, i: impl Into<SlotIndex>) -> Option<Self::Slot>;
    fn order(&self) -> usize;
    fn is_fully_self_dual(&self) -> bool;

    // Provided methods
    fn string_rep(&self) -> String { ... }
    fn external_structure(&self) -> Vec<Self::Slot> { ... }
    fn to_shell(self) -> TensorShell<Self>
       where Self: Sized { ... }
    fn contains_matching(&self, slot: &Self::Slot) -> bool { ... }
    fn external_reps(
        &self,
    ) -> Vec<Representation<<Self::Slot as IsAbstractSlot>::R>> { ... }
    fn external_indices(&self) -> Vec<<Self::Slot as IsAbstractSlot>::Aind> { ... }
    fn same_content(&self, other: &Self) -> bool { ... }
    fn match_index(&self, other: &Self) -> Option<(bool, usize, usize)> { ... }
    fn match_indices(
        &self,
        other: &Self,
    ) -> Option<(Permutation, Vec<bool>, Vec<bool>)> { ... }
    fn traces(&self) -> Vec<[usize; 2]> { ... }
    fn shape(&self) -> Vec<Dimension> { ... }
    fn reps(&self) -> Vec<Representation<<Self::Slot as IsAbstractSlot>::R>> { ... }
    fn same_external(&self, other: &Self) -> bool { ... }
    fn find_permutation(&self, other: &Self) -> Result<Permutation> { ... }
    fn strides_column_major(&self) -> Result<Vec<usize>> { ... }
    fn strides_row_major(&self) -> Result<Vec<usize>> { ... }
    fn strides(&self) -> Result<Vec<usize>> { ... }
    fn verify_indices<C: AsRef<[ConcreteIndex]>>(
        &self,
        indices: C,
    ) -> Result<()> { ... }
    fn flat_index<C: AsRef<[ConcreteIndex]>>(
        &self,
        indices: C,
    ) -> Result<FlatIndex> { ... }
    fn expanded_index(&self, flat_index: FlatIndex) -> Result<ExpandedIndex> { ... }
    fn co_expanded_index(
        &self,
        flat_index: FlatIndex,
    ) -> Result<DualConciousExpandedIndex> { ... }
    fn index_iter(&self) -> TensorStructureIndexIterator<'_, Self> 
       where Self: Sized { ... }
    fn is_scalar(&self) -> bool { ... }
    fn size(&self) -> Result<usize> { ... }

}

Required Associated Types

type Slot: IsAbstractSlot + DualSlotTo<Dual = Self::Slot>

type Indexed: TensorStructure<Indexed = Self::Indexed, Slot = Self::Slot>

Required Methods

fn reindex( self, indices: &[<Self::Slot as IsAbstractSlot>::Aind], ) -> Result<PermutedStructure<Self::Indexed>, StructureError>

fn dual(self) -> Self

fn external_structure_iter(&self) -> impl Iterator<Item = Self::Slot>

fn external_dims_iter(&self) -> impl Iterator<Item = Dimension>

fn external_reps_iter( &self, ) -> impl Iterator<Item = Representation<<Self::Slot as IsAbstractSlot>::R>>

fn external_indices_iter( &self, ) -> impl Iterator<Item = <Self::Slot as IsAbstractSlot>::Aind>

fn get_aind( &self, i: impl Into<SlotIndex>, ) -> Option<<Self::Slot as IsAbstractSlot>::Aind>

fn get_rep( &self, i: impl Into<SlotIndex>, ) -> Option<Representation<<Self::Slot as IsAbstractSlot>::R>>

fn get_dim(&self, i: impl Into<SlotIndex>) -> Option<Dimension>

fn get_slot(&self, i: impl Into<SlotIndex>) -> Option<Self::Slot>

fn order(&self) -> usize

fn is_fully_self_dual(&self) -> bool

Provided Methods

fn string_rep(&self) -> String

fn external_structure(&self) -> Vec<Self::Slot>

returns the list of slots that are the external indices of the tensor

fn to_shell(self) -> TensorShell<Self>
where Self: Sized,

fn contains_matching(&self, slot: &Self::Slot) -> bool

fn external_reps( &self, ) -> Vec<Representation<<Self::Slot as IsAbstractSlot>::R>>

fn external_indices(&self) -> Vec<<Self::Slot as IsAbstractSlot>::Aind>

fn same_content(&self, other: &Self) -> bool

checks if the tensor has the same exact structure as another tensor

fn match_index(&self, other: &Self) -> Option<(bool, usize, usize)>

Given two TensorStructures, returns the index of the first matching slot in each external index list, along with a boolean indicating if there is a single match

fn match_indices( &self, other: &Self, ) -> Option<(Permutation, Vec<bool>, Vec<bool>)>

Given two TensorStructures, returns the index of the first matching slot in each external index list

fn traces(&self) -> Vec<[usize; 2]>

Identify the repeated slots in the external index list

fn shape(&self) -> Vec<Dimension>

yields the (outwards facing) shape of the tensor as a list of dimensions

fn reps(&self) -> Vec<Representation<<Self::Slot as IsAbstractSlot>::R>>

fn same_external(&self, other: &Self) -> bool

checks if externally, the two tensors are the same

fn find_permutation(&self, other: &Self) -> Result<Permutation>

find the permutation of the external indices that would make the two tensors the same. Applying the permutation to other should make it the same as self

fn strides_column_major(&self) -> Result<Vec<usize>>

yields the strides of the tensor in column major order

fn strides_row_major(&self) -> Result<Vec<usize>>

yields the strides of the tensor in row major order

fn strides(&self) -> Result<Vec<usize>>

By default, the strides are row major

fn verify_indices<C: AsRef<[ConcreteIndex]>>(&self, indices: C) -> Result<()>

Verifies that the list of indices provided are valid for the tensor

Errors

Mismatched order = if the length of the indices is different from the order of the tensor,

Index out of bounds = if the index is out of bounds for the dimension of that index

fn flat_index<C: AsRef<[ConcreteIndex]>>(&self, indices: C) -> Result<FlatIndex>

yields the flat index of the tensor given a list of indices

Errors

Same as Self::verify_indices

fn expanded_index(&self, flat_index: FlatIndex) -> Result<ExpandedIndex>

yields the expanded index of the tensor given a flat index

Errors

Index out of bounds = if the flat index is out of bounds for the tensor

fn co_expanded_index( &self, flat_index: FlatIndex, ) -> Result<DualConciousExpandedIndex>

fn index_iter(&self) -> TensorStructureIndexIterator<'_, Self>
where Self: Sized,

yields an iterator over the indices of the tensor

fn is_scalar(&self) -> bool

if the tensor has no (external) indices, it is a scalar

fn size(&self) -> Result<usize>

yields the size of the tensor, i.e. the product of the dimensions. This is the length of the vector of the data in a dense tensor

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.