spenso::tensors::data::sparse

Struct SparseTensor

pub struct SparseTensor<T, I = OrderedStructure> {
    pub elements: HashMap<FlatIndex, T>,
    pub zero: T,
    pub structure: I,
}

Sparse data tensor, generic on storage type T, and structure type I.

Stores data in a hashmap of usize, using ahash’s hashmap. The usize key is the flattened index of the corresponding position in the dense tensor

Fields
elements: HashMap<FlatIndex, T>zero: Tstructure: I
Implementations

impl<T, S: TensorStructure> SparseTensor<T, S>

pub fn iter_trace( &self, trace_indices: [usize; 2], ) -> SparseTensorTraceIterator<'_, T, S>

Creates an iterator that traces over two indices

Arguments

trace_indices - The indices to trace over

impl<T: RefZero, S: TensorStructure> SparseTensor<T, S>

pub fn ref_zero(&self) -> T

impl<T, S: TensorStructure> SparseTensor<T, S>

pub fn map_structure<S2>(self, f: impl Fn(S) -> S2) -> SparseTensor<T, S2>
where S2: TensorStructure,

pub fn map_structure_fallible<S2, E>( self, f: impl Fn(S) -> Result<S2, E>, ) -> Result<SparseTensor<T, S2>, E>
where S2: TensorStructure,

pub fn map_data_ref<U>(&self, f: impl Fn(&T) -> U) -> SparseTensor<U, S>
where S: Clone,

pub fn map_data_ref_result<U, E>( &self, f: impl Fn(&T) -> Result<U, E>, ) -> Result<SparseTensor<U, S>, E>
where S: Clone,

pub fn map_data<U>(self, f: impl Fn(T) -> U) -> SparseTensor<U, S>

pub fn map_data_ref_mut<U>( &mut self, f: impl FnMut(&mut T) -> U, ) -> SparseTensor<U, S>
where S: Clone,

pub fn map_data_mut(&mut self, f: impl FnMut(&mut T))
where S: Clone,

impl<U, I> SparseTensor<U, I>
where I: TensorStructure + Clone,

pub fn try_upgrade<T>(&self) -> Option<Cow<'_, SparseTensor<U::LCM, I>>>
where U: TrySmallestUpgrade<T>, U::LCM: Clone,

impl<T, I> SparseTensor<T, I>
where I: TensorStructure,

pub fn empty(structure: I, zero: T) -> Self

Create a new empty sparse tensor with the given structure

pub fn is_empty_at_expanded(&self, indices: &[ConcreteIndex]) -> bool

Checks if there is a value at the given indices

pub fn is_empty_at_flat(&self, index: FlatIndex) -> bool

pub fn density(&self) -> f64

Calulates how dense the tensor is, i.e. the ratio of non-zero elements to total elements

pub fn to_dense(&self) -> DenseTensor<T, I>
where T: Clone, I: Clone,

Converts the sparse tensor to a dense tensor, with the same structure

pub fn to_dense_with(&self, zero: &T) -> DenseTensor<T, I>
where T: Clone, I: Clone,

pub fn smart_set(&mut self, indices: &[ConcreteIndex], value: T) -> Result<()>
where T: IsZero,

fallible smart set method, returns an error if the indices are out of bounds. If the value is zero, it removes the element at the given indices.

pub fn from_data( data: impl IntoIterator<Item = (Vec<ConcreteIndex>, T)>, structure: I, zero: T, ) -> Result<Self>

Generates a new sparse tensor from the given data and structure

pub fn smart_get(&self, indices: &[ConcreteIndex]) -> Result<Cow<'_, T>>
where T: Default + Clone,

fallible smart get method, returns an error if the indices are out of bounds. If the index is in the bTree return the value, else return zero.

impl<T, I> SparseTensor<T, I>
where I: Clone + TensorStructure,

pub fn convert_to<U>(&self) -> SparseTensor<U, I>
where U: for<'a> From<&'a T>,