spenso::tensors::data

Trait HasTensorData

pub trait HasTensorData: HasStructure {
    type Data: Clone;

    // Required methods
    fn data(&self) -> Vec<Self::Data>;
    fn indices(&self) -> Vec<ExpandedIndex>;
    fn hashmap(&self) -> IndexMap<ExpandedIndex, Self::Data>;
    fn symhashmap(
        &self,
        name: Symbol,
        args: &[Atom],
    ) -> HashMap<Atom, Self::Data>;
}

Trait for getting the data of a tensor

Required Associated Types

type Data: Clone

Required Methods

fn data(&self) -> Vec<Self::Data>

Returns all the data in the tensor, withouth any structure

fn indices(&self) -> Vec<ExpandedIndex>

Returns all the indices of the tensor, the order of the indices is the same as the order of the data

fn hashmap(&self) -> IndexMap<ExpandedIndex, Self::Data>

Returns a hashmap of the data, with the (expanded) indices as keys

fn symhashmap(&self, name: Symbol, args: &[Atom]) -> HashMap<Atom, Self::Data>

Returns a hashmap of the data, with the the shadowed indices as keys

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.