spenso::tensors::data

Trait SetTensorData

pub trait SetTensorData {
    type SetData;

    // Required methods
    fn set(
        &mut self,
        indices: &[ConcreteIndex],
        value: Self::SetData,
    ) -> Result<()>;
    fn set_flat(&mut self, index: FlatIndex, value: Self::SetData) -> Result<()>;
}

Trait for setting the data of a tensor

Required Associated Types

type SetData

Required Methods

fn set(&mut self, indices: &[ConcreteIndex], value: Self::SetData) -> Result<()>

Set the data at the given indices, returns an error if the indices are out of bounds

Errors

Forwards the error from TensorStructure::verify_indices

fn set_flat(&mut self, index: FlatIndex, value: Self::SetData) -> Result<()>