spenso::tensors::parametric::atomcore

Trait TensorAtomMaps

pub trait TensorAtomMaps {
    type ContainerData<T>;
    type AtomContainer;
    type Ref<'a>
       where Self: 'a;

    // Required methods
    fn replace<'b, P: Into<BorrowedOrOwned<'b, Pattern>>>(
        &self,
        pattern: P,
    ) -> ReplaceBuilderGeneric<'b, Self::Ref<'_>, Self>
       where Self: Sized;
    fn replace_multiple<T: BorrowReplacement>(
        &self,
        replacements: &[T],
    ) -> Self::AtomContainer;
    fn replace_multiple_repeat<T: BorrowReplacement>(
        &self,
        replacements: &[T],
    ) -> Self::AtomContainer;
    fn replace_multiple_mut<T: BorrowReplacement>(&mut self, replacements: &[T]);
    fn replace_multiple_repeat_mut<T: BorrowReplacement>(
        &mut self,
        replacements: &[T],
    );
    fn replace_map_mut<F: Fn(AtomView<'_>, &Context, &mut Settable<'_, Atom>)>(
        &mut self,
        m: &F,
    );
    fn coefficient<T: AtomCore>(&self, x: T) -> Self::AtomContainer;
    fn together(&self) -> Self::AtomContainer;
    fn apart(&self, x: Symbol) -> Self::AtomContainer;
    fn cancel(&self) -> Self::AtomContainer;
    fn factor(&self) -> Self::AtomContainer;
    fn collect_num(&self) -> Self::AtomContainer;
    fn expand(&self) -> Self::AtomContainer;
    fn expand_via_poly<E: Exponent, T: AtomCore>(
        &self,
        var: Option<T>,
    ) -> Self::AtomContainer;
    fn expand_in<'a, T: Into<AtomOrView<'a>>>(
        &self,
        var: T,
    ) -> Self::AtomContainer;
    fn expand_num(&self) -> Self::AtomContainer;
    fn derivative(&self, x: Symbol) -> Self::AtomContainer;
    fn series<T: AtomCore>(
        &self,
        x: Symbol,
        expansion_point: T,
        depth: Rational,
        depth_is_absolute: bool,
    ) -> Result<Self::ContainerData<Series<AtomField>>, SeriesError>;
    fn nsolve<'a, N: SingleFloat + EvaluationDomain + Real + PartialOrd, V: Into<BorrowedOrOwned<'a, Indeterminate>>>(
        &self,
        x: V,
        init: N,
        prec: N,
        max_iterations: usize,
    ) -> Result<Self::ContainerData<N>, SolveError>;
    fn evaluate<A: AtomCore + KeyLookup, T: Real + EvaluationDomain + FixedPrecision>(
        &self,
        const_map: &HashMap<A, T>,
    ) -> Result<Self::ContainerData<T>, EvaluationError>;
    fn zero_test(
        &self,
        iterations: usize,
        tolerance: f64,
    ) -> Self::ContainerData<ConditionResult>;
    fn set_coefficient_ring(
        &self,
        vars: &Arc<Vec<PolyVariable>>,
    ) -> Self::AtomContainer;
    fn to_float(&self, decimal_prec: u32) -> Self::AtomContainer;
    fn map_coefficient<F: Fn(CoefficientView<'_>) -> Coefficient + Copy>(
        &self,
        f: F,
    ) -> Self::AtomContainer;
    fn rationalize(&self, relative_error: &Rational) -> Self::AtomContainer;
    fn to_polynomial<R: EuclideanDomain + ConvertToRing, E: Exponent>(
        &self,
        field: &R,
        var_map: Option<Arc<Vec<PolyVariable>>>,
    ) -> Self::ContainerData<MultivariatePolynomial<R, E>>;
    fn to_polynomial_in_vars<E: Exponent>(
        &self,
        var_map: &Arc<Vec<PolyVariable>>,
    ) -> Self::ContainerData<MultivariatePolynomial<AtomField, E>>;
    fn to_rational_polynomial<R: EuclideanDomain + ConvertToRing, RO: EuclideanDomain + PolynomialGCD<E>, E: PositiveExponent>(
        &self,
        field: &R,
        out_field: &RO,
        var_map: Option<Arc<Vec<PolyVariable>>>,
    ) -> Self::ContainerData<RationalPolynomial<RO, E>>
       where RationalPolynomial<RO, E>: FromNumeratorAndDenominator<R, RO, E> + FromNumeratorAndDenominator<RO, RO, E>;
    fn to_factorized_rational_polynomial<R: EuclideanDomain + ConvertToRing, RO: EuclideanDomain + PolynomialGCD<E>, E: PositiveExponent>(
        &self,
        field: &R,
        out_field: &RO,
        var_map: Option<Arc<Vec<PolyVariable>>>,
    ) -> Self::ContainerData<FactorizedRationalPolynomial<RO, E>>
       where FactorizedRationalPolynomial<RO, E>: FromNumeratorAndFactorizedDenominator<R, RO, E> + FromNumeratorAndFactorizedDenominator<RO, RO, E>,
             MultivariatePolynomial<RO, E>: Factorize;
    fn to_canonical_string(&self) -> Self::ContainerData<String>;
    fn map_terms_single_core(
        &self,
        f: impl Fn(AtomView<'_>) -> Atom + Clone,
    ) -> Self::AtomContainer;
    fn map_terms(
        &self,
        f: impl Fn(AtomView<'_>) -> Atom + Send + Sync + Clone,
        n_cores: usize,
    ) -> Self::AtomContainer;
    fn to_pattern(&self) -> Self::ContainerData<Pattern>;
    fn replace_map<F: Fn(AtomView<'_>, &Context, &mut Settable<'_, Atom>)>(
        &self,
        m: &F,
    ) -> Self::AtomContainer;

}

Required Associated Types

type ContainerData<T>

type AtomContainer

type Ref<'a> where Self: 'a

Required Methods

fn replace<'b, P: Into<BorrowedOrOwned<'b, Pattern>>>( &self, pattern: P, ) -> ReplaceBuilderGeneric<'b, Self::Ref<'_>, Self>
where Self: Sized,

fn replace_multiple<T: BorrowReplacement>( &self, replacements: &[T], ) -> Self::AtomContainer

Replace all occurrences of the patterns, where replacements are tested in the order that they are given.

fn replace_multiple_repeat<T: BorrowReplacement>( &self, replacements: &[T], ) -> Self::AtomContainer

fn replace_multiple_mut<T: BorrowReplacement>(&mut self, replacements: &[T])

fn replace_multiple_repeat_mut<T: BorrowReplacement>( &mut self, replacements: &[T], )

fn replace_map_mut<F: Fn(AtomView<'_>, &Context, &mut Settable<'_, Atom>)>( &mut self, m: &F, )

fn coefficient<T: AtomCore>(&self, x: T) -> Self::AtomContainer

Collect terms involving the literal occurrence of x.

fn together(&self) -> Self::AtomContainer

Write the expression over a common denominator.

fn apart(&self, x: Symbol) -> Self::AtomContainer

Write the expression as a sum of terms with minimal denominators.

fn cancel(&self) -> Self::AtomContainer

Cancel all common factors between numerators and denominators. Any non-canceling parts of the expression will not be rewritten.

fn factor(&self) -> Self::AtomContainer

Factor the expression over the rationals.

fn collect_num(&self) -> Self::AtomContainer

Collect numerical factors by removing the numerical content from additions. For example, -2*x + 4*x^2 + 6*x^3 will be transformed into -2*(x - 2*x^2 - 3*x^3).

The first argument of the addition is normalized to a positive quantity.

fn expand(&self) -> Self::AtomContainer

Expand an expression. The function [AtomCore::expand_via_poly] may be faster.

fn expand_via_poly<E: Exponent, T: AtomCore>( &self, var: Option<T>, ) -> Self::AtomContainer

Expand the expression by converting it to a polynomial, optionally only in the indeterminate var. The parameter E should be a numerical type that fits the largest exponent in the expanded expression. Often, u8 or u16 is sufficient.

fn expand_in<'a, T: Into<AtomOrView<'a>>>(&self, var: T) -> Self::AtomContainer

Expand an expression in the variable var. The function [AtomCore::expand_via_poly] may be faster.

fn expand_num(&self) -> Self::AtomContainer

Distribute numbers in the expression, for example: 2*(x+y) -> 2*x+2*y.

fn derivative(&self, x: Symbol) -> Self::AtomContainer

Take a derivative of the expression with respect to x.

fn series<T: AtomCore>( &self, x: Symbol, expansion_point: T, depth: Rational, depth_is_absolute: bool, ) -> Result<Self::ContainerData<Series<AtomField>>, SeriesError>

Series expand in x around expansion_point to depth depth.

fn nsolve<'a, N: SingleFloat + EvaluationDomain + Real + PartialOrd, V: Into<BorrowedOrOwned<'a, Indeterminate>>>( &self, x: V, init: N, prec: N, max_iterations: usize, ) -> Result<Self::ContainerData<N>, SolveError>

Find the root of a function in x numerically over the reals using Newton’s method.

fn evaluate<A: AtomCore + KeyLookup, T: Real + EvaluationDomain + FixedPrecision>( &self, const_map: &HashMap<A, T>, ) -> Result<Self::ContainerData<T>, EvaluationError>

Evaluate a (nested) expression a single time. For repeated evaluations, use TensorAtomOps::evaluator and convert to an optimized version or generate a compiled version of your expression.

All variables and all user functions in the expression must occur in the map.

fn zero_test( &self, iterations: usize, tolerance: f64, ) -> Self::ContainerData<ConditionResult>

Check if the expression could be 0, using (potentially) numerical sampling with a given tolerance and number of iterations.

fn set_coefficient_ring( &self, vars: &Arc<Vec<PolyVariable>>, ) -> Self::AtomContainer

Set the coefficient ring to the multivariate rational polynomial with vars variables.

fn to_float(&self, decimal_prec: u32) -> Self::AtomContainer

Convert all coefficients to floats with a given precision decimal_prec``. The precision of floating point coefficients in the input will be truncated to decimal_prec`.

fn map_coefficient<F: Fn(CoefficientView<'_>) -> Coefficient + Copy>( &self, f: F, ) -> Self::AtomContainer

Map all coefficients using a given function.

fn rationalize(&self, relative_error: &Rational) -> Self::AtomContainer

Map all floating point and rational coefficients to the best rational approximation in the interval [self*(1-relative_error),self*(1+relative_error)].

fn to_polynomial<R: EuclideanDomain + ConvertToRing, E: Exponent>( &self, field: &R, var_map: Option<Arc<Vec<PolyVariable>>>, ) -> Self::ContainerData<MultivariatePolynomial<R, E>>

Convert the atom to a polynomial, optionally in the variable ordering specified by var_map. If new variables are encountered, they are added to the variable map. Similarly, non-polynomial parts are automatically defined as a new independent variable in the polynomial.

fn to_polynomial_in_vars<E: Exponent>( &self, var_map: &Arc<Vec<PolyVariable>>, ) -> Self::ContainerData<MultivariatePolynomial<AtomField, E>>

Convert the atom to a polynomial in specific variables. All other parts will be collected into the coefficient, which is a general expression.

This routine does not perform expansions.

fn to_rational_polynomial<R: EuclideanDomain + ConvertToRing, RO: EuclideanDomain + PolynomialGCD<E>, E: PositiveExponent>( &self, field: &R, out_field: &RO, var_map: Option<Arc<Vec<PolyVariable>>>, ) -> Self::ContainerData<RationalPolynomial<RO, E>>
where RationalPolynomial<RO, E>: FromNumeratorAndDenominator<R, RO, E> + FromNumeratorAndDenominator<RO, RO, E>,

Convert the atom to a rational polynomial, optionally in the variable ordering specified by var_map. If new variables are encountered, they are added to the variable map. Similarly, non-rational polynomial parts are automatically defined as a new independent variable in the rational polynomial.

fn to_factorized_rational_polynomial<R: EuclideanDomain + ConvertToRing, RO: EuclideanDomain + PolynomialGCD<E>, E: PositiveExponent>( &self, field: &R, out_field: &RO, var_map: Option<Arc<Vec<PolyVariable>>>, ) -> Self::ContainerData<FactorizedRationalPolynomial<RO, E>>
where FactorizedRationalPolynomial<RO, E>: FromNumeratorAndFactorizedDenominator<R, RO, E> + FromNumeratorAndFactorizedDenominator<RO, RO, E>, MultivariatePolynomial<RO, E>: Factorize,

Convert the atom to a rational polynomial with factorized denominators, optionally in the variable ordering specified by var_map. If new variables are encountered, they are added to the variable map. Similarly, non-rational polynomial parts are automatically defined as a new independent variable in the rational polynomial.

fn to_canonical_string(&self) -> Self::ContainerData<String>

Print the atom in a form that is unique and independent of any implementation details.

Anti-symmetric functions are not supported.

fn map_terms_single_core( &self, f: impl Fn(AtomView<'_>) -> Atom + Clone, ) -> Self::AtomContainer

Map the function f over all terms.

fn map_terms( &self, f: impl Fn(AtomView<'_>) -> Atom + Send + Sync + Clone, n_cores: usize, ) -> Self::AtomContainer

Map the function f over all terms, using parallel execution with n_cores cores.

fn to_pattern(&self) -> Self::ContainerData<Pattern>

fn replace_map<F: Fn(AtomView<'_>, &Context, &mut Settable<'_, Atom>)>( &self, m: &F, ) -> Self::AtomContainer

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".