spenso::tensors::parametric::atomcore

Trait TensorAtomOps

pub trait TensorAtomOps: HasStructure {
    // Required methods
    fn coefficient_list<E: Exponent, T: AtomCore>(
        &self,
        xs: &[T],
    ) -> Vec<(Atom, Atom)>;
    fn to_evaluation_tree(
        &self,
        fn_map: &FunctionMap,
        params: &[Atom],
    ) -> Result<EvalTreeTensor<SymComplex<Rational>, Self::Structure>, EvaluationError>;
    fn evaluator(
        &self,
        fn_map: &FunctionMap,
        params: &[Atom],
        optimization_settings: &OptimizationSettings,
    ) -> Result<EvalTensor<ExpressionEvaluator<SymComplex<Rational>>, Self::Structure>, EvaluationError>;
    fn get_all_symbols(&self, include_function_symbols: bool) -> HashSet<Symbol>;
    fn get_all_indeterminates(
        &self,
        enter_functions: bool,
    ) -> HashSet<AtomView<'_>>;
    fn contains_symbol(&self, s: Symbol) -> bool;
    fn contains<T: AtomCore>(&self, s: T) -> bool;
    fn is_polynomial(
        &self,
        allow_not_expanded: bool,
        allow_negative_powers: bool,
    ) -> Option<HashSet<AtomView<'_>>>;
}

Required Methods

fn coefficient_list<E: Exponent, T: AtomCore>( &self, xs: &[T], ) -> Vec<(Atom, Atom)>

Collect terms involving the same power of x in xs, where xs is a list of indeterminates. Return the list of key-coefficient pairs

fn to_evaluation_tree( &self, fn_map: &FunctionMap, params: &[Atom], ) -> Result<EvalTreeTensor<SymComplex<Rational>, Self::Structure>, EvaluationError>

Convert nested expressions to a tree suitable for repeated evaluations with different values for params. All variables and all user functions in the expression must occur in the map.

fn evaluator( &self, fn_map: &FunctionMap, params: &[Atom], optimization_settings: &OptimizationSettings, ) -> Result<EvalTensor<ExpressionEvaluator<SymComplex<Rational>>, Self::Structure>, EvaluationError>

Create an efficient evaluator for a (nested) expression. All free parameters must appear in params and all other variables and user functions in the expression must occur in the function map. The function map may have nested expressions.

fn get_all_symbols(&self, include_function_symbols: bool) -> HashSet<Symbol>

Get all symbols in the expression, optionally including function symbols.

fn get_all_indeterminates(&self, enter_functions: bool) -> HashSet<AtomView<'_>>

Get all variables and functions in the expression.

fn contains_symbol(&self, s: Symbol) -> bool

Returns true iff self contains the symbol s.

fn contains<T: AtomCore>(&self, s: T) -> bool

Returns true iff self contains a literally.

fn is_polynomial( &self, allow_not_expanded: bool, allow_negative_powers: bool, ) -> Option<HashSet<AtomView<'_>>>

Check if the expression can be considered a polynomial in some variables, including redefinitions. For example f(x)+y is considered a polynomial in f(x) and y, whereas f(x)+x is not a polynomial.

Rational powers or powers in variables are not rewritten, e.g. x^(2y) is not considered polynomial in x^y.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.